blazor wasm modify

+++ Logic/Scenery/SceneryModel.cs
@@ -22,10 +22,10 @@ namespace ImageCreatePlaid
         public int Alpha { get; set; }
         public int BaseBalance { get; set; }
 
-        public int BaseColorRed { get; set; }
-        public int BaseColorGreen { get; set; }
-        public int BaseColorBlue { get; set; }
-        public int BaseAlpha { get; set; }
+        public byte BaseColorRed { get; set; }
+        public byte BaseColorGreen { get; set; }
+        public byte BaseColorBlue { get; set; }
+        public byte BaseAlpha { get; set; }
 
         public string StepPattern { get; set; }
         public string StepColorPattern { get; set; }
@@ -36,10 +36,10 @@ namespace ImageCreatePlaid
         /// <summary>
         /// 縦線１
         /// </summary>
-        public int VerticalColorRed1 { get; set; }
-        public int VerticalColorGreen1 { get; set; }
-        public int VerticalColorBlue1 { get; set; }
-        public int VerticalColorAlpha1 { get; set; }
+        public byte VerticalColorRed1 { get; set; }
+        public byte VerticalColorGreen1 { get; set; }
+        public byte VerticalColorBlue1 { get; set; }
+        public byte VerticalColorAlpha1 { get; set; }
         public int VerticalSize1 { get; set; }
         public int VerticalCount1 { get; set; }
         public int VerticalDottedWeight { get; set; }
@@ -48,20 +48,20 @@ namespace ImageCreatePlaid
         /// <summary>
         /// 縦線２
         /// </summary>
-        public int VerticalColorRed2 { get; set; }
-        public int VerticalColorGreen2 { get; set; }
-        public int VerticalColorBlue2 { get; set; }
-        public int VerticalColorAlpha2 { get; set; }
+        public byte VerticalColorRed2 { get; set; }
+        public byte VerticalColorGreen2 { get; set; }
+        public byte VerticalColorBlue2 { get; set; }
+        public byte VerticalColorAlpha2 { get; set; }
         public int VerticalSize2 { get; set; }
         public int VerticalCount2 { get; set; }
 
         /// <summary>
         /// 横線１
         /// </summary>
-        public int HorizontalColorRed1 { get; set; }
-        public int HorizontalColorGreen1 { get; set; }
-        public int HorizontalColorBlue1 { get; set; }
-        public int HorizontalColorAlpha1 { get; set; }
+        public byte HorizontalColorRed1 { get; set; }
+        public byte HorizontalColorGreen1 { get; set; }
+        public byte HorizontalColorBlue1 { get; set; }
+        public byte HorizontalColorAlpha1 { get; set; }
         public int HorizontalSize1 { get; set; }
         public int HorizontalCount1 { get; set; }
         public int HorizontalDottedWeight { get; set; }
@@ -70,10 +70,10 @@ namespace ImageCreatePlaid
         /// <summary>
         /// 横線２
         /// </summary>
-        public int HorizontalColorRed2 { get; set; }
-        public int HorizontalColorGreen2 { get; set; }
-        public int HorizontalColorBlue2 { get; set; }
-        public int HorizontalColorAlpha2 { get; set; }
+        public byte HorizontalColorRed2 { get; set; }
+        public byte HorizontalColorGreen2 { get; set; }
+        public byte HorizontalColorBlue2 { get; set; }
+        public byte HorizontalColorAlpha2 { get; set; }
         public int HorizontalSize2 { get; set; }
         public int HorizontalCount2 { get; set; }
